+++ src/lang_utils/error_codes.ml
@@ -193,8 +193,6 @@ let error_codes : (string * string option) list =
     "M0213", None; (* Parenthetical note on shared functions is disallowed *)
     "M0214", None; (* Expected type of field in parenthetical note differs from inferred *)
     "M0216", None; (* Stable variable must stable subtype *)
-    "M0219", None; (* Missing `transient` *)
-    "M0220", None; (* Missing `persistent` *)
     "M0221", None; (* Failed to determine type for type pattern field *)
     "M0224", None; (* Overlapping dot resolution *)
     "M0225", None; (* A mixin cannot be used as an entry point *)