+++ src/mo_config/args.mli
@@ -6,4 +6,3 @@ val package_args : (Arg.key * Arg.spec * Arg.doc) list
 val inclusion_args : (Arg.key * Arg.spec * Arg.doc) list
 val ai_args : (Arg.key * Arg.spec * Arg.doc) list
 val migration_args : (Arg.key * Arg.spec * Arg.doc) list
-val persistent_actors_args : (Arg.key * Arg.spec * Arg.doc) list