+++ src/pipeline/test_field_srcs.ml
@@ -127,7 +127,7 @@ let run_get_sources_test source =
     ~finally:(fun () -> Mo_config.Flags.typechecker_combine_srcs := false)
 
 let%expect_test "" =
-  let s = {|persistent actor {
+  let s = {|actor {
   class Class1() = self {
     public func meth(_ : Int) : Nat {
       return 1